Add expr to fold_toggle_options if it is valid
Add expr folding when we first create and when we setup FoldParagraphs.
If we have an existing foldexpr, then we probably want to be able to
cycle to it.

It's possible that you set foldexpr, invoke ToggleFold, clear foldexpr,
and now can cycle to use a nonexistent foldexpr, but I don't think
that's a common use case worth covering. We're not checking that there
are markers defined or valid fold-syntax either.