Work around FastFold behavior
Newer versions of FastFold use the new OptionSet autocmd to detect
changing fdm=syntax and flip it to fdm=manual. We can detect if
fdm=manual and act like it was syntax (likely what we set it to). To
prevent breaks where user puts manual in their list (can't think of a
good reason), we only do this fixup if we failed to find the fdm.
However, fastfold will break this setup unless manual comes after
syntax.